feat: UserEmailsModule
