+++ project/libs/notify/user-emails/README.md
@@ -1,0 +1,7 @@
+# user-emails
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test user-emails` to execute the unit tests via [Jest](https://jestjs.io).
