+++ project/libs/notify/user-emails/src/index.ts
@@ -1,0 +1,1 @@
+export { UserEmailsModule } from './module/user-emails.module';
